requirements.txt
flask
flask-cors
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn
werkzeug